Implement AlarmScreen UI with "I'm Safe" Button
Added UI elements to the AlarmScreen to display blinking colors, an alarm emoji, and the "I'm Safe" button.
Used the widget.alertData to dynamically display the names of areas received from the alert data.
Implemented the _onImSafeButtonPressed function to stop the sound and navigate back to the previous screen when the "I'm Safe" button is pressed.